collector/requirements.txt
openmeteo-requests
requests-cache
retry-requests
pandas
pika
python-dotenv
